moved all data into separate data dir
