+++ tsconfig.json
@@ -29,7 +29,10 @@
     "app/components/blocks/navbar.tsx",
     "app/components/blocks/**/*.js",
     "app/portfolio/meme-generator/page.jsx",
-    "app/portfolio/notes-app/page.jsx"
+    "app/portfolio/notes-app/page.jsx",
+    "app/data/portfolioData.js",
+    "app/data/resumeData.js",
+    "app/data/timelineData.js"
   ],
   "exclude": ["node_modules"]
 }
